refactor: changing card-detail class with reusable popup class

+++ scripts/script.js
@@ -27,19 +27,23 @@ const initialCards = [
 ];
 
 //element
-let profile = document.querySelector(".profile");
-let popUp = document.querySelector(".popup");
-let popUpForm = popUp.querySelector(".popup__container");
-let popUpClose = document.querySelectorAll(".popup .popup__close");
-let editBtn = profile.querySelector(".profile__edit-button");
-let cards = document.querySelector(".cards");
-let cardDelete = cards.querySelectorAll(".card .card__delete");
-let cardLike = cards.querySelectorAll(".card .card__like");
-let popUpAdd = document.querySelector(".popup_add");
-let popUpAddForm = popUpAdd.querySelector(".popup__container");
-let popUpAddName = popUpAddForm.querySelector(".popup__input_name-card");
-let popUpAddLink = popUpAddForm.querySelector(".popup__input_link-card");
-let addBtn = profile.querySelector(".profile__add-button");
+const profile = document.querySelector(".profile");
+const popUp = document.querySelector(".popup");
+const popUpForm = popUp.querySelector(".popup__container");
+const popUpClose = document.querySelector(".popup__close");
+const editBtn = profile.querySelector(".profile__edit-button");
+const cards = document.querySelector(".cards");
+let popUpOpened;
+// document.querySelector(".popup_opened");
+const popUpAdd = document.querySelector(".popup_add");
+const popUpAddForm = popUpAdd.querySelector(".popup__container");
+const popUpAddName = popUpAddForm.querySelector(".popup__input_name-card");
+const popUpAddLink = popUpAddForm.querySelector(".popup__input_link-card");
+const addBtn = profile.querySelector(".profile__add-button");
+const popUpDetail = document.querySelector(".popup_detail");
+const popUpDetailForm = popUpDetail.querySelector(".popup__container");
+const detailImg = popUpDetailForm.querySelector(".popup__img");
+const detailLabel = popUpDetailForm.querySelector(".popup__label");
 
 //profile value
 let profileName = profile.querySelector(".profile__name");
@@ -49,53 +53,74 @@ let profileJob = profile.querySelector(".profile__job");
 let inputName = popUp.querySelector(".popup__input_name");
 let inputJob = popUp.querySelector(".popup__input_about");
 
+//template
 const cardTemplate = document.querySelector("#card").content;
-const cardDetailSection = document.querySelector(".card-detail-section");
-const detailTemplate = document.querySelector("#card-detail").content;
+// const cardDetailSection = document.querySelector(".card-detail-section");
+// const detailTemplate = document.querySelector("#card-detail").content;
+// let detailElement = detailTemplate
+//   .querySelector(".card-detail")
+//   .cloneNode(true);
 
-function card() {
+function loadCard() {
   initialCards.forEach((data) => {
     let cardElement = createCard(data.name, data.link);
     cards.append(cardElement);
   });
 }
 //init card
-card();
+loadCard();
 
-// init card detail
-const detailElement = detailTemplate
-  .querySelector(".card-detail")
-  .cloneNode(true);
-const detailContainer = detailElement.querySelector(".card-detail__container");
-const detailImg = detailElement.querySelector(".card-detail__img");
+// function loadCardDetail() {
+//   const detailContainer = detailElement.querySelector(
+//     ".card-detail__container"
+//   );
+//   detailImg = detailElement.querySelector(".card-detail__img");
+//   detailLabel = detailElement.querySelector(".card-detail__label");
+//   const closeBtn = detailElement.querySelector(".popup__close");
+//   closeBtn.addEventListener("click", (e) => {
+//     closePopUp(e);
+//   });
+//   detailContainer.append(detailImg);
+//   detailContainer.append(detailLabel);
+//   detailContainer.append(closeBtn);
+//   detailElement.append(detailContainer);
+//   cardDetailSection.append(detailElement);
+// }
+//init card-detail
+// loadCardDetail();
 
-const detailLabel = detailElement.querySelector(".card-detail__label");
-const closeBtn = detailElement.querySelector(".popup__close_card-detail");
-detailLabel.textContent = null;
-closeBtn.addEventListener("click", () => {
-  detailElement.classList.remove("card-detail_active");
-});
-detailContainer.append(detailImg);
-detailContainer.append(detailLabel);
-detailContainer.append(closeBtn);
-detailElement.append(detailContainer);
-cardDetailSection.append(detailElement);
-
-function popupOpen() {
-  popUp.classList.add("popup_opened");
-  let name = profileName.textContent;
-  let job = profileJob.textContent;
-  inputName.value = name;
-  inputJob.value = job;
+function OpenPopUp(e) {
+  if (e.target == editBtn) {
+    popUp.classList.add("popup_opened");
+    let name = profileName.textContent;
+    let job = profileJob.textContent;
+    inputName.value = name;
+    inputJob.value = job;
+  } else if (e.target == addBtn) {
+    popUpAdd.classList.add("popup_opened");
+  } else if (e.target.classList == "card__background") {
+    popUpDetail.classList.add("popup_opened");
+  }
+  popUpOpened = document.querySelector(".popup_opened");
+  popUpOpened.addEventListener("click", (e) => {
+    if (e.target.classList.contains("popup__close")) {
+      closePopUp(e);
+    }
+  });
 }
-function popupClose() {
-  popUp.classList.remove("popup_opened");
-  popUpAdd.classList.remove("popup_opened");
+function closePopUp(e) {
+  const element = e.target;
+  const popupContainer = element.closest(".popup__container");
+  const popup = popupContainer.closest(".popup");
+  if (popup) {
+    popup.classList.remove("popup_opened");
+  }
 }
 function handleProfileFormSubmit(evt) {
   evt.preventDefault();
   profileName.textContent = inputName.value;
   profileJob.textContent = inputJob.value;
+  closePopUp(evt);
 }
 function handleAddForm(evt) {
   evt.preventDefault();
@@ -103,13 +128,13 @@ function handleAddForm(evt) {
     name: popUpAddName.value,
     link: popUpAddLink.value,
   };
-  initialCards.unshift(newCard);
-  const name = initialCards[0].name;
-  const link = initialCards[0].link;
+  const name = newCard.name;
+  const link = newCard.link;
   const cardElement = createCard(name, link);
   cards.prepend(cardElement);
   popUpAddName.value = null;
   popUpAddLink.value = null;
+  closePopUp(evt);
 }
 function createCard(name, link) {
   const cardElement = cardTemplate.querySelector(".card").cloneNode(true);
@@ -118,20 +143,20 @@ function createCard(name, link) {
   const deleteCard = cardElement.querySelector(".card__delete");
   const likeCard = cardElement.querySelector(".card__like");
   cardImage.src = link;
-  cardImage.addEventListener("click", () => {
-    handleCardImage(name, link);
-  });
   cardTitle.textContent = name;
   cardElement.append(cardImage);
   cardElement.append(cardTitle);
   cardElement.append(deleteCard);
   cardElement.append(likeCard);
 
-  deleteCard.addEventListener("click", (e) => {
-    handleDeleteCard(e);
-  });
-  likeCard.addEventListener("click", (e) => {
-    activeLike(e);
+  cardElement.addEventListener("click", (e) => {
+    if (e.target == deleteCard) {
+      handleDeleteCard(e);
+    } else if (e.target == likeCard) {
+      activeLike(e);
+    } else if (e.target == cardImage) {
+      handleCardImage(name, link, e);
+    }
   });
 
   return cardElement;
@@ -141,42 +166,29 @@ function activeLike(e) {
   e.target.classList.toggle("card__like_active");
 }
 function handleDeleteCard(e) {
-  let parent = e.target.parentElement;
+  let parent = e.target.closest(".card");
   parent.remove();
 }
 
-function handleCardImage(name, link) {
+function handleCardImage(name, link, e) {
   detailLabel.textContent = name;
   detailImg.src = link;
-  detailElement.classList.add("card-detail_active");
-  document.body.scrollTop = 0;
-  document.documentElement.scrollTop = 0;
+  OpenPopUp(e);
+  // detailElement.classList.add("card-detail_active");
 }
 
-function addCard() {
-  popUpAdd.classList.toggle("popup_opened");
-}
-
-editBtn.addEventListener("click", function () {
-  popupOpen();
-});
-
-popUpClose.forEach((e) => {
-  e.addEventListener("click", (e) => {
-    popupClose(e);
-  });
+editBtn.addEventListener("click", (e) => {
+  OpenPopUp(e);
 });
 
 popUpForm.addEventListener("submit", function (evt) {
   handleProfileFormSubmit(evt);
-  popupClose();
 });
 
 popUpAddForm.addEventListener("submit", (evt) => {
   handleAddForm(evt);
-  popupClose();
 });
 
-addBtn.addEventListener("click", () => {
-  addCard();
+addBtn.addEventListener("click", (e) => {
+  OpenPopUp(e);
 });
